requirements.txt
python-telegram-bot==21.0.1
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
requests==2.32.3